move source_graph into parser
