+++ frb_codegen/src/lib.rs
@@ -34,7 +34,6 @@ mod markers;
 mod method_utils;
 mod others;
 mod parser;
-mod source_graph;
 mod target;
 mod transformer;
 pub mod utils;
